refactor: file 구조 변경 package file 구조 변경
