+++ backend/src/main/java/com/ssafy/trip/repository/attracion/AttractionRepository.java
@@ -1,4 +1,4 @@
-package com.ssafy.trip.repository;
+package com.ssafy.trip.repository.attracion;
 
 import com.ssafy.trip.Entity.Attraction;
 import org.springframework.data.jpa.repository.JpaRepository;
